Move to batching module
